chore: 팟캐스트 수정

See merge request s13-bigdata-recom-sub1/S13P21D101!155

+++ backend/my-issue/src/main/java/com/ssafy/myissue/podcast/domain/PodcastSubtitle.java
@@ -25,9 +25,9 @@ public class PodcastSubtitle {
     private String line;
 
     @Column(name = "start_time")
-    private double startTime;
+    private long startTime;
 
-    public static PodcastSubtitle of(Podcast podcast, int speaker, String line, double startTime) {
+    public static PodcastSubtitle of(Podcast podcast, int speaker, String line, long startTime) {
         return PodcastSubtitle.builder()
                 .podcast(podcast)
                 .speaker(speaker)
